Unified embeddings access. IndexPath only if included. Work in progress.
