+++ src/test/scala/com/johnsnowlabs/nlp/annotators/ner/crf/NerCrfApproachTestSpec.scala
@@ -1,6 +1,7 @@
 package com.johnsnowlabs.nlp.annotators.ner.crf
 
 import com.johnsnowlabs.nlp._
+import com.johnsnowlabs.nlp.embeddings.EmbeddingsHelper
 import org.scalatest.FlatSpec
 
 class NerCrfApproachTestSpec extends FlatSpec {
@@ -65,11 +66,21 @@ class NerCrfApproachTestSpec extends FlatSpec {
 
 
   it should "correctly handle entities param" in {
+
+    val smallEmbeddings = EmbeddingsHelper.loadEmbeddings(
+      "src/test/resources/random_embeddings_dim4.txt",
+      spark,
+      "TEXT",
+      4,
+      false
+    )
+
     val restrictedModel = new NerCrfModel()
       .setEntities(Array("PER", "LOC"))
       .setModel(nerModel.model.getOrDefault)
       .setOutputCol(nerModel.getOutputCol)
       .setInputCols(nerModel.getInputCols)
+      .setEmbeddings(smallEmbeddings.get)
 
     val tagged = restrictedModel.transform(nerInputDataset)
     val annotations = Annotation.collect(tagged, "ner").flatten
